Rename a few accessors to the proper names/place them in the correct package and add missing accessor entries into the json file.
