+++ src/main/java/org/spongepowered/common/mixin/accessor/entity/player/PlayerAbilitiesAccessor.java
@@ -22,28 +22,16 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.mixin.accessor.block.tileentity;
+package org.spongepowered.common.mixin.accessor.entity.player;
 
-import net.minecraft.potion.Effect;
-import net.minecraft.tileentity.BeaconTileEntity;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import net.minecraft.entity.player.PlayerAbilities;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
-import java.util.Set;
+@Mixin(PlayerAbilities.class)
+public interface PlayerAbilitiesAccessor {
 
-@Mixin(BeaconTileEntity.class)
-public interface BeaconTileEntityAccessor {
+    @Accessor("flySpeed") void accessor$setFlySpeed(float flySpeed);
 
-    @Accessor("VALID_EFFECTS") static Set<Effect> getValidEffects() {
-        throw new IllegalStateException("Untransformed accessor!");
-    }
-
-    @Accessor("primaryEffect") @Nullable Effect accessor$getPrimaryEffect();
-
-    @Accessor("primaryEffect") void accessor$setPrimaryEffect(@Nullable Effect effect);
-
-    @Accessor("secondaryEffect") @Nullable Effect accessor$getSecondaryEffect();
-
-    @Accessor("secondaryEffect") void accessor$setSecondaryEffect(@Nullable Effect effect);
+    @Accessor("walkSpeed") void accessor$setWalkSpeed(float walkSpeed);
 }
