+++ src/main/resources/mixins.common.accessor.json
@@ -9,6 +9,7 @@
     "mixins": [
         "advancements.AdvancementManagerAccessor",
         "advancements.CriteriaTriggersAccessor",
+        "block.AbstractSkullBlockAccessor",
         "block.FireBlockAccessor",
         "block.RedstoneWireBlockAccessor",
         "entity.boss.dragon.phase.PhaseTypeAccessor",
@@ -17,6 +18,7 @@
         "entity.item.FallingBlockEntityAccessor",
         "entity.item.FireworkRocketEntityAccessor",
         "entity.item.HangingEntityAccessor",
+        "entity.monster.BlazeEntityAccessor",
         "entity.monster.CreeperEntityAccessor",
         "entity.monster.EndermanEntityAccessor",
         "entity.monster.EndermiteEntityAccessor",
@@ -28,7 +30,7 @@
         "entity.passive.PigEntityAccessor",
         "entity.passive.SheepEntityAccessor",
         "entity.passive.WolfEntityAccessor",
-        "entity.player.PlayerCapabilitiesAccessor",
+        "entity.player.PlayerAbilitiesAccessor",
         "entity.player.PlayerEntityAccessor",
         "entity.projectile.AbstractArrowEntityAccessor",
         "entity.projectile.ArrowEntityAccessor",
@@ -37,8 +39,11 @@
         "entity.AgeableEntityAccessor",
         "entity.AreaEffectCloudEntityAccessor",
         "entity.EntityAccessor",
+        "entity.EntityPredicateAccessor",
         "entity.LivingEntityAccessor",
         "entity.MobEntityAccessor",
+        "inventory.container.AbstractFurnaceContainerAccessor",
+        "inventory.container.BeaconContainerAccessor",
         "inventory.container.BrewingStandContainerAccessor",
         "inventory.container.CraftingResultSlotAccessor",
         "inventory.container.ContainerAccessor",
@@ -69,8 +74,9 @@
         "scoreboard.ScoreAccessor",
         "scoreboard.ScoreboardAccessor",
         "scoreboard.ScoreObjectiveAccessor",
-        "server.management.PlayerListAccessor",
         "server.MinecraftServerAccessor",
+        "tileentity.BannerPatternAccessor",
+        "tileentity.BeaconTileEntityAccessor",
         "tileentity.BrewingStandTileEntityAccessor",
         "tileentity.CommandBlockLogicAccessor",
         "tileentity.EndGatewayTileEntityAccessor",
@@ -78,8 +84,10 @@
         "tileentity.MobSpawnerTileEntityAccessor",
         "tileentity.SkullTileEntityAccessor",
         "util.math.AxisAlignedBBAccessor",
+        "util.registry.SimpleRegistryAccessor",
         "util.text.StringTextComponentAccessor",
         "util.text.StyleAccessor",
+        "util.text.TextFormattingAccessor",
         "util.CombatEntryAccessor",
         "util.CombatTrackerAccessor",
         "util.DamageSourceAccessor",
@@ -90,6 +98,7 @@
         "world.chunk.storage.RegionFileAccessor",
         "world.dimension.DimensionTypeAccessor",
         "world.gen.feature.DungeonsFeatureAccessor",
+        "world.raid.RaidAccessor",
         "world.server.ServerWorldAccessor",
         "world.spawner.AbstractSpawnerAccessor",
         "world.storage.SaveHandlerAccessor",
@@ -101,6 +110,7 @@
         ],
     "server": [
         "server.dedicated.DedicatedServerAccessor",
+        "server.management.PlayerListAccessor",
         "server.management.UserListAccessor",
         "server.management.UserListEntryAccessor"
     ]
